FIX : Optional Points' visibility
Fixed a bug that when u switched back and forth on a shader that had an optional points his visibility came back to one even if the checkbox wasn't checked. Added a feature where optional point are now better handled has it is now not anymore the whole node's visibility that is changed but only point by point.